Created design (state machine & pseudocode), almost coherent
Do we really need a "terminated" state for threads?
Is it a good idea to use the stack for storing registers when threads
are not running?
Can we unify setSignal for interrupts and non-interrupts without making
the code much slower?
How do make invocation of thread switching more similar for
non-interrupt and interrupt cases?