Optimized thread switch
Each thread now has both a user- and a supervisor-stack. D0-A6+USP are
stored on the supervisor stack. The only statically allocated data per
task is SSP and thread state.

These changes were done to reduce the cycle count inside of the
thread-switch exception handler.

It's not entirely certain that AmigaOS likes applications which switch
supervisor stacks -- but in small tests it seems to work.

TODO: build 68010+ stackframe on supervisor stack when initializing a
thread.